+++ tests/test.c
@@ -26,7 +26,7 @@ void test_json_decode()
 {
     char jstr[] = "{\n"
     "    \"ip\": \"192.168.3.2\",\n"
-    "    \"port\": 502,\n"
+    "    \"port\": \"502\",\n"
     "    \"slave\": 22,\n"
     "    \"tid\": 1,\n"
     "    \"mode\": \"tcp\",\n"
@@ -41,7 +41,7 @@ void test_json_decode()
         printf("%s\n", cJSON_Print(json));
         printf("---------\n");
         printf("ip:%s\n", cJSON_GetObjectItem(json, "ip")->valuestring);
-        printf("port:%d\n",cJSON_GetObjectItem(json, "port")->valueint);
+        printf("port:%s\n",cJSON_GetObjectItem(json, "port")->valuestring);
         printf("mode:%s\n", cJSON_GetObjectItem(json, "mode")->valuestring);
         printf("addr:%d\n",cJSON_GetObjectItem(json, "addr")->valueint);
         
@@ -82,13 +82,13 @@ int test_init_tcp_handle_and_connect()
     LOG(enable_syslog, "Init TCP handle");
     mbtcp_handle_s *handle = NULL;
 
-    mbtcp_init_handle (&handle, "172.16.9.170", 502);
+    mbtcp_init_handle (&handle, "172.16.9.170", "502");
     //char * reason = NULL;
     //mbtcp_do_connect(handle, &reason);
 
     LOG(enable_syslog, "Get TCP handle");
     handle = NULL;
-    mbtcp_get_handle (&handle, "172.16.9.170", 502);
+    mbtcp_get_handle (&handle, "172.16.9.170", "502");
     
     LOG(enable_syslog, "Try to connect to slave");
     char * reason = NULL;
@@ -109,8 +109,8 @@ void test_multiple_add_find_zhahs()
         memset(handle, 0, sizeof(mbtcp_handle_s));
         handle->connected = false;
         strcpy(handle->key.ip, "192.168.10.12");
-        handle->key.port = idx;
-        handle->ctx = modbus_new_tcp(handle->key.ip, handle->key.port);
+        sprintf(handle->key.port, "%d", idx);       
+        handle->ctx = modbus_new_tcp_pi(handle->key.ip, handle->key.port);
         LOG(enable_syslog, "handle:%d, %p\n", idx, handle);
        
         int key = idx;        
@@ -135,8 +135,8 @@ void test_multiple_add_find()
         memset(handle, 0, sizeof(mbtcp_handle_s));
         handle->connected = false;
         strcpy(handle->key.ip, "192.168.10.12");
-        handle->key.port = idx;
-        handle->ctx = modbus_new_tcp(handle->key.ip, handle->key.port);
+        sprintf(handle->key.port, "%d", idx);
+        handle->ctx = modbus_new_tcp_pi(handle->key.ip, handle->key.port);
         HASH_ADD(hh, servers, key, sizeof(mbtcp_key_s), handle);
         LOG(enable_syslog, "handle:%d, %p\n", idx, handle);
     }
@@ -151,12 +151,12 @@ void test_multiple_add_find()
         mbtcp_handle_s query, *ptr;
         memset(&query, 0, sizeof(mbtcp_handle_s));
         strcpy(query.key.ip, "192.168.10.12");
-        query.key.port = idx;
+        sprintf(query.key.port, "%d", idx);
         HASH_FIND(hh, servers, &query.key, sizeof(mbtcp_key_s), ptr);
         
         if (ptr)
         {
-            LOG(enable_syslog, "found: %d, %s, %d, %p\n", idx, ptr->key.ip, ptr->key.port, ptr);
+            LOG(enable_syslog, "found: %d, %s, %s, %p\n", idx, ptr->key.ip, ptr->key.port, ptr);
         }
         else
         {
@@ -178,11 +178,11 @@ void test_single_zhash()
     memset(h1, 0, sizeof(h1));
     h1->connected = false;
     strcpy(h1->key.ip, "192.168.10.1");
-    h1->key.port = 555;
-    h1->ctx = modbus_new_tcp(h1->key.ip, h1->key.port);
+    strcpy(h1->key.port, "555");
+    h1->ctx = modbus_new_tcp_pi(h1->key.ip, h1->key.port);
     //HASH_ADD(hh, servers, key, sizeof(mbtcp_key_s), h1);
     
-    mbtcp_key_s key = {"192.168.10.1", 555};
+    mbtcp_key_s key = {"192.168.10.1", "555"};
     
     zhashx_t *hash = zhashx_new ();
     int rc;
@@ -192,7 +192,7 @@ void test_single_zhash()
     
     //look up  
     mbtcp_handle_s * item = (mbtcp_handle_s *) zhashx_lookup (hash, &key);
-    printf("%s, %d, %d\n", item->key.ip, item->key.port, item->connected);
+    printf("%s, %s, %d\n", item->key.ip, item->key.port, item->connected);
     
     zhashx_destroy (&hash);
 }
@@ -212,8 +212,8 @@ void test_single_add_find()
     memset(h1, 0, sizeof(h1));
     h1->connected = false;
     strcpy(h1->key.ip, "192.168.10.1");
-    h1->key.port = 555;
-    h1->ctx = modbus_new_tcp(h1->key.ip, h1->key.port);
+    strcpy(h1->key.port, "555");
+    h1->ctx = modbus_new_tcp_pi(h1->key.ip, h1->key.port);
     HASH_ADD(hh, servers, key, sizeof(mbtcp_key_s), h1);
     
     // server #2
@@ -221,19 +221,19 @@ void test_single_add_find()
     memset(h2, 0, sizeof(h2));
     h2->connected = false;
     strcpy(h2->key.ip, "192.168.10.2");
-    h2->key.port = 556;
-    h2->ctx = modbus_new_tcp(h2->key.ip, h2->key.port);
+    strcpy(h2->key.port, "556");
+    h2->ctx = modbus_new_tcp_pi(h2->key.ip, h2->key.port);
     HASH_ADD(hh, servers, key, sizeof(mbtcp_key_s), h2);
     
     // find #1
     memset(&ff, 0, sizeof(mbtcp_handle_s));
     strcpy(ff.key.ip, "192.168.10.1");
-    ff.key.port = 555;
+    strcpy(ff.key.port, "555");
     HASH_FIND(hh, servers, &ff.key, sizeof(mbtcp_key_s), p);
     
     if (p)
     {
-        LOG(enable_syslog, "Found %s, %d\n", p->key.ip, p->key.port);
+        LOG(enable_syslog, "Found %s, %s\n", p->key.ip, p->key.port);
     }
     else
     {
@@ -243,12 +243,12 @@ void test_single_add_find()
     // find #2
     memset(&ff, 0, sizeof(mbtcp_handle_s));
     strcpy(ff.key.ip, "192.168.10.2");
-    ff.key.port = 556;
+    strcpy(ff.key.port, "556");
     HASH_FIND(hh, servers, &ff.key, sizeof(mbtcp_key_s), p);
     
     if (p)
     {
-        LOG(enable_syslog, "Found %s, %d\n", p->key.ip, p->key.port);
+        LOG(enable_syslog, "Found %s, %s\n", p->key.ip, p->key.port);
     }
     else
     {