+++ HISTORY.md
@@ -16,7 +16,9 @@
 - read config
 - default config
 - implement dummy modbus server for testing
-
+- Ipv4, v6 support
+- rewrite int port to char * port
 
 ## TODO List
-- daemon pid
+- daemon pid, service
+- support change timeout command